# tests/contracts/openapi-snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zoofy Automation",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/subscriptions": {
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Create Subscription Invoice",
        "description": "n8n calls this once per parsed email. Upsert on `factuur` so a re-processed email\n(e.g. label didn't stick, workflow re-run) never creates a duplicate row — n8n doesn't\nneed to check for existing rows itself before posting.",
        "operationId": "create_subscription_invoice_api_subscriptions_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionInvoiceIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionInvoiceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "List Subscription Invoices",
        "operationId": "list_subscription_invoices_api_subscriptions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionInvoiceOut"
                  },
                  "title": "Response List Subscription Invoices Api Subscriptions Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/subscriptions/{factuur}/pdf": {
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Upload Subscription Pdf",
        "operationId": "upload_subscription_pdf_api_subscriptions__factuur__pdf_post",
        "parameters": [
          {
            "name": "factuur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Factuur"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_subscription_pdf_api_subscriptions__factuur__pdf_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionInvoiceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_subscription_pdf_api_subscriptions__factuur__pdf_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_subscription_pdf_api_subscriptions__factuur__pdf_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubscriptionInvoiceIn": {
        "properties": {
          "factuur": {
            "type": "string",
            "title": "Factuur"
          },
          "kenmerk": {
            "type": "string",
            "title": "Kenmerk"
          },
          "factuurdatum": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factuurdatum"
          },
          "vervaldatum": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vervaldatum"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "thread_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Link"
          }
        },
        "type": "object",
        "required": [
          "factuur",
          "kenmerk"
        ],
        "title": "SubscriptionInvoiceIn"
      },
      "SubscriptionInvoiceOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "factuur": {
            "type": "string",
            "title": "Factuur"
          },
          "kenmerk": {
            "type": "string",
            "title": "Kenmerk"
          },
          "factuurdatum": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factuurdatum"
          },
          "vervaldatum": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vervaldatum"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "thread_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Link"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "factuur",
          "kenmerk",
          "factuurdatum",
          "vervaldatum",
          "amount",
          "thread_link",
          "pdf_url",
          "created_at"
        ],
        "title": "SubscriptionInvoiceOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}